Changed the data types of constant fields.

+++ src/main/java/com/api/management/leave/leavemanagementapi/dto/LeaveCreditsEarnedDto.java
@@ -5,12 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class LeaveCreditsEarnedDto {
-    private double daysPresent;
-    private double leaveWithoutPay;
-    private double LeaveCreditsEarned;
+    private BigDecimal daysPresent;
+    private BigDecimal leaveWithoutPay;
+    private BigDecimal LeaveCreditsEarned;
 }
